Test2/requirements.txt
streamlit==1.29.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
chromadb==0.4.18
scikit-learn==1.3.2
rank-bm25==0.2.2
accelerate==1.7.0
